requirements.txt
biopython
requests
selenium
streamlit
openai
llama-index
nltk
pypdf
chromadb